+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Mapbox welcomes participation and contributions from everyone.
 
 ### Features ✨ and improvements 🏁
 * Promote Geofencing APIs to stable. Remove `@_spi(Experimental)` from Geofencing APIs.
+* Fix an issue where the location indicator could show incorrect direction in landscape orientation.
 
 ## 11.17.0-rc.2 - 21 November, 2025